requirements.txt
clean-text
emoji
jieba
tqdm
unidecode
pyahocorasick
ijson
nltk
p_tqdm